Allow bindings to be setup by the component's tagData
This makes it so that the templating engine can provide a
`setupBindings` function on the component's `tagData` which will be used
to get the initial values from the element that will be used to
construct the View Model.

This is a 3-step fix. This initial commit makes it so that the
templating engine can provide a `setupBindings` function, but we still
fallback to can-stache-bindings if it doesn't exist. Step 2 will be to
update can-stache to provide it's own bindings using setupBindings. Step
3 will be to remove the dependency on can-stache-bindings from
can-component.