Add support for rendering components in stache templates
This adds the can.viewInsert symbol to each component instance so when a component is rendered as a value in a stache template, it can be correctly removed later.

The tests cover this specific functionality and the overall integration of changes made in can-view-live and can-stache to make this feature work.

Part of https://github.com/canjs/can-stache/issues/502