Fix bug where connectedCallback gets called twice for `new Components({ ... })`
When a component is programmatically instantiated and its element is inserted into the DOM, two callbacks run:

- can-view-callbacks.tag (because it listens for the custom element to be inserted into the DOM) https://github.com/canjs/can-component/blob/b60dbf6c4fda12318eab4c7f513fb3081a4122e6/can-component.js#L324-L326
- can-dom-mutate.onNodeInsertion (because the element wasn’t in the DOM when it was created, but does get inserted later) https://github.com/canjs/can-component/blob/b60dbf6c4fda12318eab4c7f513fb3081a4122e6/can-component.js#L578-L581

This caused connectedCallback to be called twice.

This commit fixes that by checking in the can-view-callbacks.tag callback whether the component’s element has a symbol that’s added when the component’s constructor function runs.